+++ resources/user_keys.rb
@@ -25,6 +25,6 @@ default_action :create
 attribute :name, :kind_of => String, :name_attribute => true
 attribute :user_databag, :kind_of => String, :default => 'users'
 attribute :user_query, :kind_of => String, :default => '*:*'
-attribute :key_dir, :kind_of => String, :default => '/etc/openvpn/keys'
+attribute :base_dir, :kind_of => String, :default => '/etc/openvpn'
 attribute :port, :kind_of => Integer, :default => 1194
 attribute :instance, :kind_of => String, :default => nil