+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG for opsline-openvpn
 
+## 0.1.3
+* integrated multidaemon support, tls key generation, and chef-managed persistence of user and server keys into default recipe
+* added iptables-based firewall restrictions configurable per openvpn daemon
+
 ## 0.1.2
 * adding multidaemon support